Update Git Connector
